Initial commit of GeoTrellis docs
For now this is a demo of an mdoc built site and
will house our initial effort to provide compiled
docs for GeoTrellis examples.

It was decided to make this a separate repo while
we evaluate the tech and build up an initial set
of resources. Our intention, if successful, is to
move these docs into the locationtech/geotrellis
repo with the next major GeoTrellis release.

It remains unclear where we would serve these docs.